fastapi/requirements.txt
pandas==2.2.3
uvicorn==0.32.1
gunicorn==23.0.0
numpy==2.2.0
fastapi==0.115.6
mysql-connector-python==9.1.0
